# openapi.yaml
openapi: 3.1.0
info:
  title: Predictify API
  version: 0.0.1
  description: Backend API for Predictify — a Stellar/Soroban prediction-markets dApp
servers:
  - url: /
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            requestId:
              type: string
          required:
            - code
      required:
        - error
    ValidationErrorBody:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            details:
              nullable: true
          required:
            - code
      required:
        - error
...
    Market:
      type: object
      properties:
        id:
          type: string
        question:
          type: string
        status:
          type: string
        metadata:
          nullable: true
        version:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - question
        - status
        - version
        - createdAt
...
    PatchMarketRequest:
      type: object
      properties:
        question:
          type: string
        metadata:
          nullable: true
        expectedVersion:
          type: integer
          minimum: 0
      required:
        - expectedVersion
...
    NotificationCategory:
      type: string
      enum:
        - market_resolved
        - claim_ready
        - dispute_opened
    NotificationChannel:
      type: string
      enum:
        - email
        - webhook
    NotificationPreference:
      type: object
      properties:
        category:
          $ref: '#/components/schemas/NotificationCategory'
        channel:
          $ref: '#/components/schemas/NotificationChannel'
        enabled:
          type: boolean
      required:
        - category
        - channel
        - enabled
    NotificationPreferencesResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            preferences:
              type: array
              items:
                $ref: '#/components/schemas/NotificationPreference'
          required:
            - preferences
      required:
        - data
    PatchNotificationPreferencesRequest:
      type: object
      properties:
        preferences:
          type: array
          items:
            $ref: '#/components/schemas/NotificationPreference'
          minItems: 1
      required:
        - preferences
    MarkNotificationsReadResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            updatedCount:
              type: integer
              minimum: 0
          required:
            - updatedCount
      required:
        - data
    NotificationId:
      type: string
      format: uuid
    MarkNotificationsReadRequest:
      type: object
      properties:
        notificationIds:
          type: array
          items:
            $ref: '#/components/schemas/NotificationId'
        markAllAsRead:
          type: boolean
      additionalProperties: false
...
    ReconciliationSidePosition:
      type: object
      properties:
        stellarAddress:
          type: string
        outcome:
          type: string
        amount:
          type: string
      required:
        - stellarAddress
        - outcome
        - amount
    ReconciliationSummary:
      type: object
      properties:
        totalKeys:
          type: integer
        matches:
          type: integer
        mismatches:
          type: integer
        missingOnChain:
          type: integer
        missingInDb:
          type: integer
      required:
        - totalKeys
        - matches
        - mismatches
        - missingOnChain
        - missingInDb
    ReconciliationDiffEntry:
      type: object
      properties:
        key:
          type: object
          properties:
            stellarAddress:
              type: string
            outcome:
              type: string
          required:
            - stellarAddress
            - outcome
        dbAmount:
          type: string
        onChainAmount:
          type: string
          nullable: true
        difference:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - match
            - mismatch
            - missing_on_chain
            - missing_in_db
      required:
        - key
        - dbAmount
        - onChainAmount
        - difference
        - status
    MarketReconciliationResult:
      type: object
      properties:
        marketId:
          type: string
        correlationId:
          type: string
        generatedAt:
          type: string
          format: date-time
        status:
          type: string
          enum:
            - ok
            - partial
        dbSnapshot:
          type: object
          properties:
            positions:
              type: array
              items:
                $ref: '#/components/schemas/ReconciliationSidePosition'
            totalAmount:
              type: string
          required:
            - positions
            - totalAmount
        onChainSnapshot:
          type: object
          properties:
            positions:
              type: array
              items:
                $ref: '#/components/schemas/ReconciliationSidePosition'
            totalAmount:
              type: string
            available:
              type: boolean
            source:
              type: string
            unavailableReason:
              type: string
              nullable: true
          required:
            - positions
            - totalAmount
            - available
            - source
            - unavailableReason
        summary:
          $ref: '#/components/schemas/ReconciliationSummary'
        diffs:
          type: array
          items:
            $ref: '#/components/schemas/ReconciliationDiffEntry'
      required:
        - marketId
        - correlationId
        - generatedAt
        - status
        - dbSnapshot
        - onChainSnapshot
        - summary
        - diffs
...
paths:
...
  /api/markets/{id}:
    get:
      operationId: getMarketById
      tags:
        - Markets
      summary: Get a market by ID
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      responses:
        '200':
          description: Market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
                required:
                  - data
              examples:
                default:
                  value:
                    data:
                      id: market-001
                      question: Will the US win the 2026 FIFA World Cup?
                      status: active
                      metadata:
                        category: sports
                        resolutionSource: official
                      version: 1
                      createdAt: '2026-01-10T12:00:00.000Z'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    patch:
      operationId: updateMarket
      tags:
        - Markets
      summary: Update a market (admin only)
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
          required: true
          name: id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchMarketRequest'
            examples:
              updateQuestion:
                value:
                  question: Will the US win the 2026 FIFA World Cup Final?
                  metadata:
                    category: sports
                  expectedVersion: 1
      responses:
        '200':
          description: Updated market
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Market'
                required:
                  - data
              examples:
                updatedMarket:
                  value:
                    data:
                      id: market-001
                      question: Will the US win the 2026 FIFA World Cup Final?
                      status: active
                      metadata:
                        category: sports
                      version: 2
                      createdAt: '2026-01-10T12:00:00.000Z'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
              examples:
                invalidBody:
                  value:
                    error:
                      code: validation_error
                      details:
                        - path:
                            - expectedVersion
                          message: expectedVersion is required
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                notFound:
                  value:
                    error:
                      code: not_found
                      requestId: req_abc123
        '409':
          description: Version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
              examples:
                versionConflict:
                  value:
                    error:
                      code: conflict
                      requestId: req_abc123
...
  /api/notifications/preferences:
    get:
      operationId: getNotificationPreferences
      tags:
        - Notifications
      summary: Get the authenticated user’s notification preferences
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
    patch:
      operationId: patchNotificationPreferences
      tags:
        - Notifications
      summary: Update notification preferences for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchNotificationPreferencesRequest'
      responses:
        '200':
          description: Updated notification preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationPreferencesResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
  /api/notifications/mark-read:
    post:
      operationId: markNotificationsRead
      tags:
        - Notifications
      summary: Mark notifications as read for the authenticated user
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkNotificationsReadRequest'
      responses:
        '200':
          description: Notifications marked as read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkNotificationsReadResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...
  /api/admin/recon/markets/{id}:
    get:
      operationId: adminReconcileMarket
      tags:
        - Admin
      summary: On-demand market reconciliation (admin only)
      description: >-
        Compares confirmed on-chain positions against the database snapshot for a single market and returns a structured
        diff. Every call is audit-logged as `admin.reconciliation.market.inspect`. Returns `status: "partial"` when the
        on-chain adapter is not yet wired.
      security:
        - bearerAuth: []
      parameters:
        - schema:
            type: string
            minLength: 1
            maxLength: 255
            description: Market ID
          required: true
          description: Market ID
          name: id
          in: path
      responses:
        '200':
          description: Reconciliation result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/MarketReconciliationResult'
                required:
                  - data
        '400':
          description: Validation error — invalid market ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorBody'
        '403':
          description: Forbidden — missing or non-admin JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '404':
          description: Market not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorBody'
...